Change HentProduktPris return type to Double
Updated the HentProduktPris type alias to return a Double instead of an Int. This ensures more accurate representation of product prices, accommodating decimal values.

+++ src/main/kotlin/PlasserBestillingWorkflow/BestillingPublicTypes.kt
@@ -46,7 +46,7 @@ typealias Valideringsfeil = String
 // Dependencies
 typealias SjekkProduktKodeEksisterer = (String) -> Boolean
 typealias SjekkAdresseEksisterer = (IkkeValidertBestilling.IkkeValidertAdresse) -> Boolean
-typealias HentProduktPris = (produktkode: Produktkode) -> Int
+typealias HentProduktPris = (produktkode: Produktkode) -> Double
 typealias LagBekreftelsesEpostHtml = (PrisetBestilling) -> HtmlString
 typealias SendBekreftelsesEpost = (email: String, letter: HtmlString) -> SendEpostResultat
 typealias SjekkEpostValideringsStatus = (email: String) -> Boolean
